Move sensor to models
